refactor: [SIW-1938] modularize IT Wallet machines
