+++ apps/main-app/ts/features/itwallet/machine/eid/actions.ts
@@ -10,6 +10,7 @@ import { isRouteInNavigationState } from "../../../../utils/navigation";
 import { checkCurrentSession } from "../../../authentication/common/store/actions";
 import {
   trackItWalletIDMethodSelected,
+  trackItWalletIntroScreen,
   trackItwDeactivated,
   trackItwIdAuthenticationCompleted,
   trackItwIdVerifiedDocument,
@@ -351,6 +352,12 @@ export const createEidIssuanceActionsImplementation = (
         )
       )
     );
+  },
+
+  trackIntroScreen: ({
+    context
+  }: ActionArgs<Context, EidIssuanceEvents, EidIssuanceEvents>) => {
+    trackItWalletIntroScreen(context.level === "l3" ? "L3" : "L2");
   },
 
   trackWalletInstanceCreation: ({
