+++ apps/main-app/ts/features/itwallet/machine/upgrade/machine.ts
@@ -1,73 +1,9 @@
-import { assign, fromCallback, fromPromise, setup } from "xstate";
+import { assign } from "xstate";
 
-import { ItwSessionExpiredError } from "../../api/client";
-import {
-  LoadContextOutput,
-  RequestAccessTokenOutput,
-  RequestAccessTokenParams,
-  UpgradeCredentialOutput,
-  UpgradeCredentialParams
-} from "./actors";
-import { Context, getInitialContext } from "./context";
-import { CredentialUpgradeEvents } from "./events";
-import { mapUpgradeEventToFailure } from "./failure";
-import { Input } from "./input";
-import { Output } from "./output";
+import { getInitialContext } from "./context";
+import { itwUpgradeSetup } from "./setup";
 
-const notImplemented = () => {
-  throw new Error("Not implemented");
-};
-
-export const itwCredentialUpgradeMachine = setup({
-  types: {
-    events: {} as CredentialUpgradeEvents,
-    context: {} as Context,
-    input: {} as Input,
-    output: {} as Output
-  },
-  actions: {
-    storeCredential: notImplemented,
-    pickNextCredential: assign({
-      credentialIndex: ({ context }) => context.credentialIndex + 1
-    }),
-    setFailedCredential: assign({
-      failedCredentials: ({ context, event }) => {
-        const current = context.credentials[context.credentialIndex];
-
-        const failedEvent = mapUpgradeEventToFailure(event);
-
-        const failedCredential = {
-          ...current,
-          failure: {
-            type: failedEvent.type,
-            reason: failedEvent.reason
-          }
-        };
-
-        return [...context.failedCredentials, failedCredential];
-      }
-    }),
-    handleSessionExpired: notImplemented
-  },
-  actors: {
-    requestAccessToken: fromPromise<
-      RequestAccessTokenOutput,
-      RequestAccessTokenParams
-    >(notImplemented),
-    loadContext: fromPromise<LoadContextOutput>(notImplemented),
-    upgradeCredential: fromPromise<
-      UpgradeCredentialOutput,
-      UpgradeCredentialParams
-    >(notImplemented),
-    waitForSessionRefresh: fromCallback(notImplemented)
-  },
-  guards: {
-    isSessionExpired: ({ event }) =>
-      "error" in event && event.error instanceof ItwSessionExpiredError,
-    hasMoreCredentials: ({ context }) =>
-      context.credentialIndex < context.credentials.length - 1
-  }
-}).createMachine({
+export const itwCredentialUpgradeMachine = itwUpgradeSetup.createMachine({
   /** @xstate-layout N4IgpgJg5mDOIC5QEsAuB3AwgJ0mAdqsgIYA2AqgA5TbERgCyxAxgBbL5gB0mrYzAaw5QAxAG0ADAF1EoSgHtYaZPPyyQAD0QA2ACwBOLgHZ92gBx6AzPssSjZgDQgAnogCMAVkvGATNo8eFmYS5mb6PgC+EU5oWLj0hCQU1LT0TGwc3Lz8QviiYm4ySCAKSkSq6loIlnpcEmYe4dpGHm76Nk6uCGZuXAbtYWESdroSHlExGDh4iWRUNHSMLOycXPOpYNMJRGQiEKrcHABu8gKHU-EEO8kLacuZaymLW1dJCMfyzMTl+JJSf+pSsoKsUqrYfFx9EZ-IE3D4fG4JLoPJ1EPCJJCvHCjLojG5bJYzBMQLEXrMbht0ituOtnpdySIwNhsPJsFxKKRvgAzVkAWy4pPp11pdwyqxFmyFbw+Xx+fwBxSBP0qiHBkOhAR68MRyNRCGxfXaULc5iM9jaRii0RA+Hk9HgxUFM2FT1F1MBimBalBiAAtNo9f66sMQ6GQwjiU7tkkJVSHtlBMIPWUVN7QFUPH5IQiLIEfEYQj5HC53K11f4LHYfJZGuZIxdnTHXUsxTTm2TrsmvSqELofHq3GauLZcR49OY-JYfLp63FG3Nm3HVph5LyOWBUJAu8qfQh7Lo+mZdDVkfoTRI3LoBx4MUY-K0zJZB24jzOrUA */
   id: "itwCredentialUpgradeMachine",
   context: ({ input }) => getInitialContext(input),
