+++ apps/main-app/ts/features/itwallet/presentation/proximity/machine/machine.ts
@@ -1,99 +1,12 @@
-import {
-  and,
-  assign,
-  fromCallback,
-  fromPromise,
-  not,
-  or,
-  setup,
-  stateIn
-} from "xstate";
+import { not } from "xstate";
 
-import {
-  ProximityCommunicationLogicInput,
-  SendDocumentsActorInput,
-  SendDocumentsActorOutput,
-  SendErrorResponseActorOutput,
-  StartEngagementActorInput
-} from "./actors";
-import { Context, InitialContext } from "./context";
-import { ProximityEvents } from "./events";
-import { mapEventToFailure, ProximityFailureType } from "./failure";
-import { ItwPresentationTags } from "./tags";
+import { InitialContext } from "./context";
+import { itwProximityMachineSetup } from "./setup";
+import { bluetoothState } from "./state/bluetooth";
+import { nfcState } from "./state/nfc";
+import { presentmentState } from "./state/presentment";
 
-const notImplemented = () => {
-  throw new Error("Not implemented");
-};
-
-export const itwProximityMachine = setup({
-  types: {
-    context: {} as Context,
-    events: {} as ProximityEvents
-  },
-  actions: {
-    onInit: notImplemented,
-    /**
-     * Context manipulation
-     */
-
-    setFailure: assign(({ event }) => ({ failure: mapEventToFailure(event) })),
-    markSessionTerminated: assign(() => ({ sessionTerminated: true })),
-    resetSessionTerminated: assign(() => ({ sessionTerminated: false })),
-
-    /**
-     * Navigation
-     */
-
-    navigateToBluetoothPermissionsScreen: notImplemented,
-    navigateToBluetoothActivationScreen: notImplemented,
-    navigateToNfcActivationScreen: notImplemented,
-    navigateToPresentmentScreen: notImplemented,
-    navigateToNfcPresentmentScreen: notImplemented,
-    navigateToFailureScreen: notImplemented,
-    navigateToClaimsDisclosureScreen: notImplemented,
-    navigateToStoreconsentScreen: notImplemented,
-    navigateToSuccessScreen: notImplemented,
-    closeProximity: notImplemented,
-
-    /**
-     * Consents
-     */
-
-    grantConsent: notImplemented,
-    storeConsent: notImplemented,
-
-    /**
-     * Analytics
-     */
-
-    trackProximityStart: notImplemented,
-    trackQrCodeLoadingFailure: notImplemented
-  },
-  actors: {
-    checkBluetoothPermissions: fromPromise<boolean>(notImplemented),
-    checkBluetoothActivation: fromPromise<boolean>(notImplemented),
-    checkNfcActivation: fromPromise<boolean>(notImplemented),
-    proximityCommunicationLogic: fromCallback<
-      ProximityEvents,
-      ProximityCommunicationLogicInput
-    >(notImplemented),
-    startEngagement: fromPromise<void, StartEngagementActorInput>(
-      notImplemented
-    ),
-    sendDocuments: fromPromise<
-      SendDocumentsActorOutput,
-      SendDocumentsActorInput
-    >(notImplemented),
-    terminateSession: fromPromise<SendErrorResponseActorOutput>(notImplemented)
-  },
-  guards: {
-    hasFailure: ({ context }) => !!context.failure,
-    isNfcRetrieval: ({ context }) => context.retrievalMethod === "nfc",
-    isNfcEngagement: ({ context }) => context.engagementMode === "nfc",
-    hasTerminatedSession: ({ context }) => context.sessionTerminated,
-    hasGrantedConsent: notImplemented
-  }
-}).createMachine({
+export const itwProximityMachine = itwProximityMachineSetup.createMachine({
   id: "itwProximityMachine",
   context: { ...InitialContext },
   initial: "Idle",
@@ -107,430 +20,9 @@ export const itwProximityMachine = setup({
         }
       }
     },
-    Bluetooth: {
-      tags: [ItwPresentationTags.Loading],
-      description: "Bluetooth permission and activation gate",
-      initial: "CheckPermissions",
-      states: {
-        CheckPermissions: {
-          description: "Check if Bluetooth permissions are granted",
-          invoke: {
-            src: "checkBluetoothPermissions",
-            onDone: [
-              {
-                guard: ({ event }) => event.output,
-                target: "CheckActivation"
-              },
-              {
-                target: "RequirePermissions"
-              }
-            ],
-            onError: {
-              target: "RequirePermissions"
-            }
-          }
-        },
-        RequirePermissions: {
-          description: "Prompt the user to grant Bluetooth permissions",
-          entry: "navigateToBluetoothPermissionsScreen",
-          on: {
-            close: {
-              actions: "closeProximity"
-            },
-            continue: {
-              target: "CheckActivation"
-            }
-          }
-        },
-        CheckActivation: {
-          description: "Check if Bluetooth is enabled",
-          invoke: {
-            src: "checkBluetoothActivation",
-            onDone: [
-              {
-                guard: ({ event }) => event.output,
-                target: "Completed"
-              },
-              {
-                target: "RequireActivation"
-              }
-            ],
-            onError: {
-              target: "RequireActivation"
-            }
-          }
-        },
-        RequireActivation: {
-          description: "Prompt the user to enable Bluetooth",
-          entry: "navigateToBluetoothActivationScreen",
-          on: {
-            close: {
-              actions: "closeProximity"
-            },
-            continue: {
-              target: "Completed"
-            }
-          }
-        },
-        Completed: {
-          description: "Bluetooth gate cleared",
-          type: "final"
-        }
-      },
-      onDone: {
-        target: "#itwProximityMachine.Presentment",
-        actions: "navigateToPresentmentScreen"
-      }
-    },
-    Nfc: {
-      description: "NFC activation gate, entered only when the user opts in",
-      initial: "CheckActivation",
-      states: {
-        CheckActivation: {
-          description: "Check if NFC is enabled",
-          invoke: {
-            src: "checkNfcActivation",
-            onDone: [
-              {
-                guard: ({ event }) => event.output,
-                target: "Completed"
-              },
-              {
-                guard: ({ event }) => !event.output,
-                target: "RequireActivation"
-              }
-            ],
-            onError: {
-              target: "RequireActivation"
-            }
-          }
-        },
-        RequireActivation: {
-          description: "Prompt the user to enable NFC",
-          entry: "navigateToNfcActivationScreen",
-          on: {
-            close: {
-              // Back to the QR engagement still in progress, without committing to NFC
-              target: "#itwProximityMachine.Presentment"
-            },
-            continue: {
-              target: "Completed"
-            }
-          }
-        },
-        Completed: {
-          description: "NFC gate cleared",
-          type: "final"
-        }
-      },
-      onDone: {
-        // External transition to Presentment fully restarts proximityCommunicationLogic
-        // and startEngagement so the native session runs with the NFC configuration
-        target: "#itwProximityMachine.Presentment",
-        actions: [
-          assign({ engagementMode: "nfc" }),
-          "navigateToNfcPresentmentScreen"
-        ]
-      }
-    },
-    Presentment: {
-      description:
-        "Proximity communication lifecycle with the verifier, driven by context.engagementMode",
-      initial: "Starting",
-      invoke: {
-        id: "proximityCommunicationLogic",
-        src: "proximityCommunicationLogic",
-        input: ({ context }) => ({
-          credentials: context.credentials
-        }),
-        onError: {
-          actions: "setFailure",
-          target: "#itwProximityMachine.Failure"
-        }
-      },
-      on: {
-        "qr-code-string": {
-          target: "Presentment.AwaitingConnection",
-          actions: assign(({ event }) => ({
-            qrCodeString: event.payload
-          }))
-        },
-        "device-connecting": {
-          target: "Presentment.Connecting"
-        },
-        "device-connected": {
-          target: "Presentment.Connected",
-          actions: "trackProximityStart"
-        },
-        "device-document-request-received": [
-          {
-            // NFC engagement also starts BLE retrieval. A late BLE copy of the
-            // same request during consent teardown would re-enter
-            // EvaluatingConsent and invoke terminateSession again, double-resuming
-            // IOWalletProximity's NFC CheckedContinuation (fatal SIGTRAP).
-            guard: or([
-              stateIn("Presentment.TerminatingForConsent"),
-              stateIn("Presentment.ClaimsDisclosure"),
-              stateIn("Presentment.StoreConsent"),
-              stateIn("Presentment.SendingDocuments"),
-              stateIn("Presentment.Terminating")
-            ])
-          },
-          {
-            actions: assign(({ event }) => ({
-              proximityDetails: event.proximityDetails,
-              verifierRequest: event.verifierRequest,
-              retrievalMethod: event.retrievalMethod
-            })),
-            target: "Presentment.EvaluatingConsent"
-          }
-        ],
-        "device-disconnected": [
-          {
-            // END (0x02) flag received AFTER sendDocuments: verification complete
-            guard: stateIn("Presentment.SendingDocuments"),
-            target: "#itwProximityMachine.Success"
-          },
-          {
-            // Expected disconnect after intentional session termination for NFC
-            // retrieval — both while tearing the session down (TerminatingForConsent)
-            // and while the user reviews the request (ClaimsDisclosure).
-            guard: and([
-              or([
-                stateIn("Presentment.TerminatingForConsent"),
-                stateIn("Presentment.ClaimsDisclosure")
-              ]),
-              "isNfcRetrieval"
-            ])
-          },
-          {
-            // END (0x02) flag received BEFORE sendDocuments: verifier aborted
-            actions: "setFailure",
-            target: "Presentment.Terminating"
-          }
-        ],
-        "device-error": [
-          {
-            // Expected error during intentional session termination for NFC
-            // retrieval — consumed without failure, matching device-disconnected.
-            guard: and([
-              or([
-                stateIn("Presentment.TerminatingForConsent"),
-                stateIn("Presentment.ClaimsDisclosure")
-              ]),
-              "isNfcRetrieval"
-            ])
-          },
-          {
-            guard: not(stateIn("Presentment.Terminating")),
-            actions: "setFailure",
-            target: "#itwProximityMachine.Failure"
-          }
-        ]
-      },
-      states: {
-        Retrying: {
-          description: "Clear the failure and restart the engagement",
-          tags: [ItwPresentationTags.Loading],
-          always: {
-            target: "Starting",
-            actions: assign(() => ({ failure: undefined }))
-          }
-        },
-        Starting: {
-          description: "Start the native engagement session",
-          tags: [ItwPresentationTags.Loading],
-          entry: "resetSessionTerminated",
-          always: {
-            guard: "isNfcRetrieval",
-            actions: "navigateToNfcPresentmentScreen"
-          },
-          invoke: {
-            src: "startEngagement",
-            input: ({ context }) => ({
-              engagementMode: context.engagementMode
-            }),
-            onDone: {
-              target: "AwaitingConnection"
-            },
-            onError: [
-              {
-                guard: "isNfcRetrieval",
-                actions: "setFailure",
-                target: "#itwProximityMachine.Failure"
-              },
-              {
-                actions: ["setFailure", "trackQrCodeLoadingFailure"]
-              }
-            ]
-          },
-          on: {
-            retry: {
-              target: "Retrying"
-            }
-          }
-        },
-        AwaitingConnection: {
-          description:
-            "Engagement is live, waiting for the verifier to connect",
-          tags: [ItwPresentationTags.Presenting],
-          on: {
-            "start-nfc-presentment": {
-              target: "#itwProximityMachine.Nfc"
-            },
-            "nfc-stopped": {
-              // NFC session has ended (HCE modal closed)
-              actions: "closeProximity"
-            },
-            retry: {
-              target: "Retrying"
-            },
-            close: {
-              actions: "closeProximity"
-            }
-          }
-        },
-        Connecting: {
-          description: "Verifier is initiating the connection",
-          tags: [ItwPresentationTags.Loading],
-          always: {
-            // Pre-navigate to the (loading) claims screen for QR engagement only.
-            guard: not("isNfcEngagement"),
-            actions: "navigateToClaimsDisclosureScreen"
-          },
-          on: {
-            // NFC session has ended (HCE modal closed)
-            "nfc-stopped": "Terminating"
-          }
-        },
-        Connected: {
-          description: "Verifier connected, waiting for the document request",
-          tags: [ItwPresentationTags.Loading],
-          on: {
-            // In case of connection timeout, allows the user to exit the flow
-            close: {
-              actions: "closeProximity",
-              target: "#itwProximityMachine.Idle"
-            }
-          }
-        },
-        EvaluatingConsent: {
-          description:
-            "Decide whether to surface the consent screen or skip it",
-          always: [
-            {
-              // NFC retrieval re-enters this state after consent was already granted earlier in the session
-              guard: and(["hasGrantedConsent", "isNfcRetrieval"]),
-              target: "#itwProximityMachine.Presentment.SendingDocuments"
-            },
-            {
-              // NFC retrieval, consent not yet granted: the NFC link cannot be held
-              // open while the user reviews the request, so tear the session down
-              // (and release the SDK) before asking for consent.
-              guard: "isNfcRetrieval",
-              target: "#itwProximityMachine.Presentment.TerminatingForConsent"
-            },
-            {
-              target: "#itwProximityMachine.Presentment.ClaimsDisclosure"
-            }
-          ]
-        },
-        ClaimsDisclosure: {
-          description: "Display the requested claims for review",
-          entry: "navigateToClaimsDisclosureScreen",
-          on: {
-            "holder-consent": [
-              {
-                // NFC retrieval: restart the engagement so the verifier re-issues the request;
-                // EvaluatingConsent will then skip this screen and go straight to SendingDocuments
-                guard: "isNfcRetrieval",
-                actions: "grantConsent",
-                target: "#itwProximityMachine.Presentment.StoreConsent"
-              },
-              {
-                target: "#itwProximityMachine.Presentment.SendingDocuments"
-              }
-            ],
-            close: {
-              actions: assign(() => ({
-                failure: {
-                  type: ProximityFailureType.CONSENT_DENIED,
-                  reason: undefined
-                }
-              })),
-              target: "#itwProximityMachine.Failure"
-            }
-          }
-        },
-        TerminatingForConsent: {
-          description:
-            "NFC retrieval: terminate the live session and release the SDK before asking for consent, without leaving the proximity flow",
-          tags: [ItwPresentationTags.Loading],
-          entry: "markSessionTerminated",
-          invoke: {
-            id: "terminateSession",
-            src: "terminateSession",
-            onDone: {
-              target: "#itwProximityMachine.Presentment.ClaimsDisclosure"
-            },
-            onError: {
-              // Ignore termination failures: proceed to consent and rely on the
-              // restart's startEngagement to reset the native session.
-              target: "#itwProximityMachine.Presentment.ClaimsDisclosure"
-            }
-          }
-        },
-        StoreConsent: {
-          description:
-            "Asks user if he want to save the consent for future requests",
-          entry: "navigateToStoreconsentScreen",
-          on: {
-            "store-consent": {
-              actions: "storeConsent",
-              target: "#itwProximityMachine.Presentment.Retrying"
-            },
-            continue: {
-              target: "#itwProximityMachine.Presentment.Retrying"
-            }
-          }
-        },
-        SendingDocuments: {
-          description: "Send the accepted documents to the verifier",
-          tags: [ItwPresentationTags.Loading, ItwPresentationTags.Sending],
-          invoke: {
-            id: "sendDocuments",
-            src: "sendDocuments",
-            input: ({ context }) => ({
-              credentials: context.credentials,
-              verifierRequest: context.verifierRequest
-            }),
-            onDone: {
-              // Verifier does not acknowledge the response; completion arrives via device-disconnected
-            },
-            onError: {
-              actions: "setFailure",
-              target: "#itwProximityMachine.Failure"
-            }
-          }
-        },
-        Terminating: {
-          tags: [ItwPresentationTags.Loading],
-          description: "Send the session-termination signal to the verifier",
-          entry: "markSessionTerminated",
-          invoke: {
-            id: "terminateSession",
-            src: "terminateSession",
-            onDone: {
-              actions: "closeProximity"
-            },
-            onError: {
-              // We ignore any failure on purpose and consider presentation terminated
-              actions: "closeProximity"
-            }
-          }
-        }
-      }
-    },
+    Bluetooth: bluetoothState,
+    Nfc: nfcState,
+    Presentment: presentmentState,
     Success: {
       description: "Documents successfully sent to the verifier",
       always: {
@@ -548,44 +40,20 @@ export const itwProximityMachine = setup({
     Failure: {
       description: "An error occurred, captured in context.failure",
       entry: "navigateToFailureScreen",
-      initial: "EnsureTerminated",
-      states: {
-        EnsureTerminated: {
-          always: [
-            {
-              // NFC consent teardown already sent SESSION_TERMINATED.
-              // A second native sendErrorResponse double-resumes the
-              // IOWalletProximity CheckedContinuation (fatal SIGTRAP).
-              guard: "hasTerminatedSession",
-              target: "Idle"
-            },
-            {
-              target: "Terminating"
-            }
-          ]
+      invoke: {
+        id: "terminateSession",
+        src: "terminateSession",
+        onDone: {
+          // Attempt termination ignoring result
         },
-        Terminating: {
-          invoke: {
-            id: "terminateSession",
-            src: "terminateSession",
-            onDone: {
-              actions: "markSessionTerminated",
-              target: "Idle"
-            },
-            onError: {
-              actions: "markSessionTerminated",
-              target: "Idle"
-            }
-          }
-        },
-        Idle: {
-          description: "Termination attempted or skipped for this engagement"
+        onError: {
+          // Attempt termination ignoring any failure
         }
       },
       on: {
         close: {
           actions: "closeProximity",
-          target: "#itwProximityMachine.Idle"
+          target: "Idle"
         }
       }
     }
