+++ apps/main-app/ts/features/itwallet/trustmark/machine/machine.ts
@@ -1,87 +1,10 @@
-import { addSeconds, differenceInSeconds, isPast } from "date-fns";
-import { assign, fromPromise, not, setup } from "xstate";
+import { assign, not } from "xstate";
 
-import { trackItwTrustmarkRenewFailure } from "../../analytics";
-import { getMixPanelCredential } from "../../analytics/utils";
 import { ItwTags } from "../../machine/tags";
-import {
-  GetCredentialTrustmarkUrlActorInput,
-  GetCredentialTrustmarkUrlActorOutput,
-  GetWalletAttestationActorOutput
-} from "./actors";
-import { Context } from "./context";
-import { TrustmarkEvents } from "./events";
-import { mapEventToFailure } from "./failure";
-import { Input } from "./input";
+import { itwTrustmarkMachineSetup } from "./setup";
+import { displayingTrustmarkState } from "./state/displayingTrustmark";
 
-const notImplemented = () => {
-  throw new Error("Not implemented");
-};
-
-/**
- * Amount in seconds to wait before retrying
- */
-const backoffTimeAmounts = [1, 10, 60, 180];
-
-export const itwTrustmarkMachine = setup({
-  types: {
-    context: {} as Context,
-    input: {} as Input,
-    events: {} as TrustmarkEvents
-  },
-  actions: {
-    onInit: notImplemented,
-    storeWalletInstanceAttestation: notImplemented,
-    handleSessionExpired: notImplemented,
-    updateExpirationSeconds: assign(({ context }) => ({
-      expirationSeconds: context.expirationDate
-        ? differenceInSeconds(context.expirationDate, new Date())
-        : undefined
-    })),
-    resetTrustmark: assign({
-      trustmarkUrl: undefined,
-      expirationDate: undefined,
-      expirationSeconds: undefined
-    }),
-    incrementAttempts: assign(({ context }) => {
-      const attempts = context.attempts ? context.attempts + 1 : 1;
-      const backoffTime = backoffTimeAmounts[attempts - 1] || 180;
-      return {
-        attempts,
-        nextAttemptAt: addSeconds(new Date(), backoffTime)
-      };
-    }),
-    resetAttempts: assign({
-      attempts: undefined,
-      nextAttemptAt: undefined
-    }),
-    setFailure: assign({
-      failure: ({ event }) => mapEventToFailure(event)
-    }),
-    showRetryFailureToast: notImplemented,
-    trackTrustmarkFailure: ({ context }) => {
-      trackItwTrustmarkRenewFailure(
-        getMixPanelCredential(context.credentialType, false)
-      );
-    }
-  },
-  actors: {
-    getWalletAttestationActor:
-      fromPromise<GetWalletAttestationActorOutput>(notImplemented),
-    getCredentialTrustmarkActor: fromPromise<
-      GetCredentialTrustmarkUrlActorOutput,
-      GetCredentialTrustmarkUrlActorInput
-    >(notImplemented)
-  },
-  guards: {
-    isTrustmarkExpired: ({ context }) =>
-      context.expirationDate ? isPast(context.expirationDate) : true,
-    isSessionExpired: notImplemented,
-    hasValidWalletInstanceAttestation: notImplemented,
-    hasBackoffTimePassed: ({ context }) =>
-      context.nextAttemptAt ? isPast(context.nextAttemptAt) : true
-  }
-}).createMachine({
+export const itwTrustmarkMachine = itwTrustmarkMachineSetup.createMachine({
   id: "itwTrustmarkMachine",
   context: ({ input }) => ({
     credentialType: input.credentialType
@@ -153,33 +76,7 @@ export const itwTrustmarkMachine = setup({
         }
       }
     },
-    DisplayingTrustmark: {
-      description: "Displays the QR Code and checks if it has expired",
-      initial: "Idle",
-      entry: "resetAttempts",
-      states: {
-        Idle: {
-          after: {
-            1000: {
-              target: "Checking"
-            }
-          }
-        },
-        Checking: {
-          entry: "updateExpirationSeconds",
-          always: [
-            {
-              guard: "isTrustmarkExpired",
-              actions: "resetTrustmark",
-              target: "#itwTrustmarkMachine.CheckingWalletInstanceAttestation"
-            },
-            {
-              target: "Idle"
-            }
-          ]
-        }
-      }
-    },
+    DisplayingTrustmark: displayingTrustmarkState,
     Failure: {
       description: "This state is reached when an error occurs",
       entry: ["incrementAttempts", "trackTrustmarkFailure"],
